Telephony: Incoming Call in Background (2/3)
Patch ported from 4.2 which was done by our lovely HCP. A
very popular one so we bring it back and adapt it to 4.4

Allows user to set that Incoming Call UI will remain in background
if screen is on. Instead, a notification is added to StatusBar, which
can be used to bring up the Incoming Call UI. Ringtone and vibration
still occur as per system settings.

This has the same effect as apps marked as "immersive". It prevents
the incoming call from backgrounding the current activity.

Change-Id: If3b8debf80c6b66343cf4257f87d8b70d727c1f6
PS: reorder settings a bit

Conflicts:
	src/com/android/phone/CallFeaturesSetting.java